Refactor ChainIndexError and ChainIndexLog
