+++ playground-common/src/PSGenerator/Common.hs
@@ -36,7 +36,7 @@ import           Ledger.TimeSlot                           (SlotConfig, SlotConv
 import           Ledger.Typed.Tx                           (ConnectionError, WrongOutTypeError)
 import           Ledger.Value                              (AssetClass, CurrencySymbol, TokenName, Value)
 import           Playground.Types                          (ContractCall, FunctionSchema, KnownCurrency)
-import           Plutus.ChainIndex.Emulator.Handlers       (ChainIndexError, ChainIndexLog)
+import           Plutus.ChainIndex                         (ChainIndexError, ChainIndexLog)
 import           Plutus.ChainIndex.Tx                      (ChainIndexTx, ChainIndexTxOutputs)
 import           Plutus.ChainIndex.Types                   (BlockNumber, Depth, Page, PageSize, Point, Tip, TxStatus,
                                                             TxValidity)
